feat: create a component for Leave Org
